requirements.txt
Flask
Flask-Cors
boto3
requests
python-dotenv